refactor: extract chunk_store_refs_api.c
Move branch/tag/remote/tracking CRUD and refs blob serialize/install
into a dedicated TU. Wire PROLLY_OBJS, DOLTLITE_EXTRA_TSRC, amalgamation
engine list, and lint presence/cap (≤1200).
